Refactoring - aggregator-related analysis functions
